move comparative pie widget ui to comparative folder
