+++ packages/react-ui/src/index.js
@@ -24,7 +24,7 @@ import PolygonIcon from './assets/PolygonIcon';
 import RectangleIcon from './assets/RectangleIcon';
 import LassoIcon from './assets/LassoIcon';
 import CircleIcon from './assets/CircleIcon';
-import ComparativePieWidgetUI from './widgets/ComparativePieWidgetUI';
+import ComparativePieWidgetUI from './widgets/comparative/ComparativePieWidgetUI';
 
 const featureSelectionIcons = {
   CursorIcon,
